pubg_stat_tracker/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
Django==3.1.5
django-appconf==1.0.4
django-background-tasks==1.2.5
django-compat==1.0.15
django-crispy-forms==1.10.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.4
mysql-connector-python==8.0.22
numpy==1.20.1
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
